+++ docs/changelog.rst
@@ -3,5 +3,7 @@ Changelog
 
 `Next Release`_
 ---------------
+- Add :func:`coercion.stringify`
+- Add :func:`coercion.normalize_collection`
 
 .. _Next Release: https://github.com/dave-shawley/coercion/compare/0.0.0...master